==> analyses/molecular-subtyping-HGG/03-HGG-molecular-subtyping-cnv.Rmd <==
---
title: "High-grade Glioma Molecular Subtyping - Focal and Broad Copy Number Alterations"
author: "Chante Bethell, Stephanie J. Spielman, and Jaclyn Taroni for ALSF CCDL"
date: "2020"
output:
  html_notebook:
    toc: yes
    toc_float: yes
---

This notebook prepares focal and broad copy number alteration data for the 
purpose of subtyping HGG samples 
([`AlexsLemonade/OpenPBTA-analysis#249`](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/249)).

## Usage

This notebook is intended to be run via the command line from the top directory
of the repository as follows:

`Rscript -e "rmarkdown::render('analyses/molecular-subtyping-HGG/03-HGG-molecular-subtyping-cnv.Rmd', clean = TRUE)"`

## Set Up

```{r}
library(tidyverse)
```

### Directories

```{r}
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
# File path to results directory
input_dir <-
  file.path(root_dir, "analyses", "molecular-subtyping-HGG", "hgg-subset")
# File path to results directory
results_dir <-
  file.path(root_dir, "analyses", "molecular-subtyping-HGG", "results")

if (!dir.exists(results_dir)) {
  dir.create(results_dir)
}
```

### Read in Files

We will need the clinical/histologies file for included samples to distinguish between missing samples or samples that had only neutral copy number calls.

```{r message=FALSE}
hgg_metadata_df <- read_tsv(file.path(input_dir,
                                     "hgg_metadata.tsv"), 
                           guess_max = 1000)

```


#### Read in CN Data

```{r}
# Read in HGG subset focal CN data
cn_df <- read_tsv(file.path(input_dir, "hgg_focal_cn.tsv.gz"))

```

### Output File

```{r}
output_file <- file.path(results_dir, "HGG_cleaned_cnv.tsv")
```

## Annotated consensus CN data

We are using consensus copy number data data that has been annotated via the [`focal-cn-file-preparation`](../focal-cn-file-preparation/) module.
Note that a gene may have evidence for two kinds of copy number alteration (e.g., both _loss_ and _gain_) in these data, but no longer see instances of this when using the consensus data for these samples.

```{r}
# Select only the ID information, for use in joining below
cn_df_ids <- cn_df %>%  
  select(sample_id,
         Kids_First_Biospecimen_ID,
         Kids_First_Participant_ID)

# Filter focal CN data for PDGFRA, PTEN and MYCN status
cn_df_genes <- cn_df %>%
  rename(gene_symbol = hgnc_symbol) %>%
  dplyr::filter(gene_symbol %in% c("PDGFRA", "PTEN", "MYCN", "EGFR"))

# Summarize the data to have a single row per sample where _all_ kinds of copy number alterations are present.
cn_df_genes_status <- cn_df_genes %>% 
  group_by(gene_symbol, sample_id, Kids_First_Biospecimen_ID) %>% 
  summarize(full_status = paste(status, collapse=", ")) %>% 
  spread(gene_symbol, full_status) %>% 
  right_join(cn_df_ids) %>%
  distinct() %>%
  replace_na(list(#PDGFRA = "neutral", 
                  PTEN = "neutral", 
                  # no MYCN mutation in Hope cohort
                  #MYCN = "neutral",
                  EGFR = "neutral")) %>% 
  # Rename columns to indicate focal status explicitly
  dplyr::rename(#MYCN_focal_status   = MYCN, 
                 #PDGFRA_focal_status = PDGFRA,
                 PTEN_focal_status   = PTEN,
                EGFR_focal_status = EGFR) %>%
  # Arrange order of columns
  dplyr::select(sample_id,
                Kids_First_Biospecimen_ID, 
                Kids_First_Participant_ID, 
                everything())

head(cn_df_genes_status, n = 10)
```


```{r}
final_hgg_cn_df <- cn_df_genes_status

```

`final_hgg_cn_df` includes all the samples that are in the consensus CNV file.
So any missing CNA values at this stage come from the fact that neutral values are _not_ included in the annotated copy number files.
For WGS samples, we know that the NA is from neutral but for WXS samples, the NA is due to the fact that we did not run GISTIC - hence we need to treat them differently.

```{r}
# Find WGS samples and all the NAs can be replaced to neutral
WGS_bsid <- hgg_metadata_df %>% 
  dplyr::filter(experimental_strategy == "WGS") %>%
  pull(Kids_First_Biospecimen_ID) %>% 
  unique()

final_hgg_cn_df_wgs <- cn_df_genes_status %>%
  filter(Kids_First_Biospecimen_ID %in% WGS_bsid) %>% 
  replace(is.na(.), "neutral") %>%
  dplyr::arrange(Kids_First_Participant_ID, sample_id)

# For the remaining samples, only replace NA for focal cn calls
final_hgg_cn_df_wxs <- cn_df_genes_status %>%
  filter(!Kids_First_Biospecimen_ID %in% WGS_bsid) %>% 
  dplyr::mutate(#MYCN_focal_status = replace_na(MYCN_focal_status, "neutral"), 
                #PDGFRA_focal_status = replace_na(PDGFRA_focal_status, "neutral"),
                PTEN_focal_status = replace_na(PTEN_focal_status, "neutral"),
                EGFR_focal_status = replace_na(EGFR_focal_status, "neutral"))

final_hgg_cn_df <- bind_rows(final_hgg_cn_df_wgs, final_hgg_cn_df_wxs)
```

Write this to file.

```{r}
write_tsv(final_hgg_cn_df, output_file)
```

## Session Info

```{r}
sessionInfo()
```
